VEOSS-851| changed vat owed to be big decimal

+++ app/models/financialdata/Payment.scala
@@ -22,7 +22,7 @@ import play.api.libs.json.{Format, Json}
 import java.time.LocalDate
 
 case class Payment(period: Period,
-                   amountOwed: Long,
+                   amountOwed: BigDecimal,
                    dateDue: LocalDate,
                    paymentStatus: PaymentStatus
                   )
